+++ syntheseus/reaction_prediction/inference/__init__.py
@@ -6,11 +6,17 @@ from syntheseus.reaction_prediction.inference.megan import MEGANModel
 from syntheseus.reaction_prediction.inference.mhnreact import MHNreactModel
 from syntheseus.reaction_prediction.inference.retro_knn import RetroKNNModel
 from syntheseus.reaction_prediction.inference.root_aligned import RootAlignedModel
+from syntheseus.reaction_prediction.inference.toy_models import (
+    LinearMoleculesToyModel,
+    ListOfReactionsToyModel,
+)
 
 __all__ = [
     "ChemformerModel",
     "GLNModel",
     "Graph2EditsModel",
+    "LinearMoleculesToyModel",
+    "ListOfReactionsToyModel",
     "LocalRetroModel",
     "MEGANModel",
     "MHNreactModel",