+++ syntheseus/reaction_prediction/inference/toy_models.py
@@ -3,25 +3,25 @@ from __future__ import annotations
 from typing import Sequence
 
 from syntheseus.interface.bag import Bag
+from syntheseus.interface.models import BackwardReactionModel
 from syntheseus.interface.molecule import Molecule
 from syntheseus.interface.reaction import SingleProductReaction
-from syntheseus.search.reaction_models.base import BackwardReactionModel
 
 
-class ListOfReactionsModel(BackwardReactionModel):
+class ListOfReactionsToyModel(BackwardReactionModel):
     """A model which returns reactions from a pre-defined list."""
 
     def __init__(self, reaction_list: Sequence[SingleProductReaction], **kwargs) -> None:
         super().__init__(**kwargs)
         self.reaction_list = list(reaction_list)
 
-    def _get_backward_reactions(
-        self, mols: list[Molecule]
+    def _get_reactions(
+        self, inputs: list[Molecule], num_results: int
     ) -> list[Sequence[SingleProductReaction]]:
-        return [[r for r in self.reaction_list if r.product == mol] for mol in mols]
+        return [[r for r in self.reaction_list if r.product == mol] for mol in inputs]
 
 
-class LinearMolecules(BackwardReactionModel):
+class LinearMoleculesToyModel(BackwardReactionModel):
     """
     A simple toy model of "reactions" on linear "ball-and-stick" molecules,
     where the possible reactions involve string cuts and substitutions.
@@ -82,7 +82,7 @@ class LinearMolecules(BackwardReactionModel):
 
         return output
 
-    def _get_backward_reactions(
-        self, mols: list[Molecule]
+    def _get_reactions(
+        self, inputs: list[Molecule], num_results: int
     ) -> list[Sequence[SingleProductReaction]]:
-        return [self._get_single_backward_reactions(mol) for mol in mols]
+        return [self._get_single_backward_reactions(mol) for mol in inputs]