+++ syntheseus/search/reaction_models/__init__.py
@@ -1,3 +1,0 @@
-from syntheseus.search.reaction_models.base import BackwardReactionModel
-
-__all__ = ["BackwardReactionModel"]